Определение_ориентации_справки(DonorSearch.org)/requirements.txt
anyio == 4.6.0
fastapi      ==     0.115.0
numpy        <     2.0
opencv-python ==    4.10.0.84
pillow    ==       10.4.0
pip        ==       23.2.1
pydantic    ==      2.9.2
torch       ==      2.2.2
torchvision  ==     0.17.2
uvicorn     ==      0.31.0
python-multipart
